TASK-3: Moved each car type class into a separate file
